Projeto de Testes Santander
- Projeto de Testes Santander
- Versão do Android é a Marshmallow 6.0 – 6.0.1	23;
- Testes Unitários Locais utilizando JUnit4 e Robolectric;
- Testes Instrumentados utilizando Espresso;
- Linguagem utilizada no App foi o Java;
- Consumo do WebService foi realizado através do Retrofit 2.0;
- Acesso ao banco de dados realizado com classes especialistas;
- Arquitetura aplicada foi a Clean Code;

+++ app/src/main/java/br/com/fernandodutra/testeandroidv2/dao/UserAccountDAO.java
@@ -25,6 +25,6 @@ public interface UserAccountDAO<T> {
 
     ArrayList<HMAux> findList() throws SQLException;
 
-    long nextID() throws SQLException;
+    int nextID() throws SQLException;
 
 }
